[Policy] Add SFTP upload support
Adds support for uploading via SFTP. This is done via pexpect calling
the system's locally available SFTP binary. If either that binary or
pexpect are unavailable on the local system, we will exit gracefully and
report the issue. This allows sos to keep python3-pexpect as a
recommends rather than a hard dependency.

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>