[report] Fix display of default values in --list for magic values
When plugin options use a magic value as a placeholder for logic checks
used during collections, those magic values can be confusing to users
when reported via `--list`.

This was originally limited to the timeout-related options, and was
mitigated by a note in the option description. However, this approach is
not maintainable as more plugin options implement magic values to allow
complex determination of what the actual value should be during an
execution - for example, with the new `--namespaces` option and plugin
options allowing users to override that global value.

Fix the former by implementing a `TIMEOUT_DEFAULT` constant. Fix the
latter by reassigning the internal magic value from `None` to the
implied value of `0`.

Related: #2547

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>