[plugins] Add containers enablement trigger
Adds a `containers` enablement trigger for plugins that currently check
for specifically named containers as part of their `setup()`
collections.

This is primarily the openstack plugins checking for their respective
component api's to be running in a container. In these cases the name
pattern used in the `container_exists()` checks are used for the
enablement trigger as well.

Resolves: #2561

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>