[cirrus] Update avocado commands with new test directories
With the addition of the `cleaner_tests/` and `collect_tests/` subdirs,
update the CirrusCI config to include these directories, while still
avoiding the `product_tests/` directory for the "base" stageone and
stagetwo tasks.

Additionally, install python3-pexpect for the stagetwo tests. Currently,
this is only utilized to test the use of the placeholder component for
`collect`. However, in the future we'll be doing actual `collect` tests
which would require this to be installed anyways.

Resolves: #2573

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>