[Plugin] Add default journal collection for Plugin services
Adds automatic collection of journals for any service defined in a
plugin's `services` tuple, if that service exists on the system.

Updates several plugins to define a `services` tuple in place of
manually calling `add_journal()` (and as a bonus, removing redundant
`add_service_status()` calls that are already performed based on the
tuple).

Additionally, add an appropriate tag to each `add_journal` call for
easier collection lookup in the manifest.

Resolves: #2579

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>