[cleaner] Iterate over matches with most precise match first
When matching strings in parsers to do obfuscation, we should be using
the most precise matches found first, rather than matching in the order
a match is hit. This ensures that we correctly obfuscate an entire
string, rather than potentially only partial substring(s) that exist
within the entire match.

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>