[tests] use Avocado "recursive" behavior by default
Avocado will, by default, use the "recursive" behavior when looking
for tests.  It means that the class hierarchy will be crawled
recursively, until, for "avocado-instrumented" tests, the top-most
parent "avocado.Test" is found.

When the "enable" behavior is activated, it forces the class to be
considered one containing avocado-instrumented tests, but, it disables
the recursive behavior and only the tests local to that specific class
are found.

Signed-off-by: Cleber Rosa <crosa@redhat.com>