[tests] Work around for Avocado's safeloader issue
The Avocado method of finding tests is one in which tests are not
executed.  While most users will also want to run tests, and not only
find them, the process of listing tests ("avocado list") does not
eval/run any code.

This makes it harder, when finding tests, to always behave like
evaluated code.  One example is a current Avocado bug in which imports
that may contain classes with tests, are not found if they come from
a compound, absolute import.  While looking for modules with classes,
Avocado does not currently look into every single entry in Python's
import path (sys.path/PYTHONPATH).

The workaround is to use relative imports, so that Avocado will be
able to precisely find and look at the file containing the interesting
class.  In this case, it helps Avocado find the next class in the
hierarchy, and then find all the relevant tests.

Reference: https://github.com/avocado-framework/avocado/issues/4625
Signed-off-by: Cleber Rosa <crosa@redhat.com>