[cleaner] Add error message when using empty file
This patche solves the situation where we attempt to
use an empty json file for cleaner matches. Before this
it issues the following error:

Could not initialize 'clean': Expecting value: line 1
column 1 (char 0)

With the patch, it prints an error message similar to
the one we use when there's no mapping file.

Resolves: RHBZ#1929172

Signed-off-by: Jose Castillo <jcastillo@redhat.com>