kata_containers: Update to the 2.x release
Kata Containers 1.x, which is the base for this plugin, has reached its
EOL on May 12th 2021*.

Based on that, let's update the sos plugin to match the current Fedora
RPM (kata-containers, instead of kata-runtime).

Currently, kata-containers can only be used in the kubernetes world and
it'll always be spawn either by CRI-O or containerd, and then it'll log
to the journal using their identifier, and the responsibility of
collecting the CRI runtime journal is left for the CRI runtimes.

*: https://medium.com/kata-containers/about-kata-containers-releases-8b3d91d1166e

Signed-off-by: Fabiano Fidêncio <fidencio@redhat.com>