[build,tests] Include .po files for manual builds, add cirrus test
First, adds a MANIFEST.in file for distutils to reference, so that we
can easily ensure that the .po files are included in any manual build
from the git checkout.

Second, add a cirrus task to ensure that basic rpm builds from the git
checkout can be done by the user.

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>