[lgtm] Suppress false positive alerts
Adds an alert suppression for 2 LGTM.com alerts. First, one involving
the overriding of `load_options()` when we're initializing the
placeholder component for `collect` on systems that do not have the
proper support for it.

Second, an alert about logging sensitive data when we are reporting the
temporary anonymous user used to leverage the RH SFTP server when the
user does not provide credentials.

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>