[sos] Read config section for clean when `--clean` is used
When a `report` or `collect` run would use `--clean` or `--mask` to do
in-line obfuscation of collected reports, sos would not read the config
section for clean - it would only be read if `sos clean` was called
directly. As such, users would need to manually specify config file
values for each run.

Alleviate this gap by reading the config section for `clean` if either
of the cleaner options are used. Do this before we apply cmdline options
so that we maintain our order of precedence.

Related: RHBZ#1950350

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>