[cirrus] Upload logs from failed tests to GCE storage bucket
Adds failure handling to cirrus tasks that actually run sos so that if
an error is encountered in the test suite, the logs from the tests are
uploaded to the GCE cloud storage bucket associated with the GCE sos
project.

This will also make those logs available from the cirrus task page.

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>