[unpackaged] Limit package_manager file list to relevant paths
Since the `unpackaged` plugin looks for all files within $PATH, we
should in turn limit the list of files returned by the package manager
that we're comparing against to the same paths.

Otherwise, we are iterating over a needlessly large list of packaged
files outside of the filesystem paths that we actually care about.

On systems with a large number of installed packages, this can result in
a packaged file list that is smaller by tens or even hundreds of
thousands of entries and in turn results in a much faster `setup()` time
for the plugin.

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>