[options] Fix logging on plugopts in effective sos command
First, provide a special-case handling for plugin options specified in
sos.conf in `SoSOptions.to_args().has_value()` that allows for plugin
options to be included in the "effective options now" log message.

Second, move the logging of said message (and thus the merging of
preset options, if used), to being _prior_ to the loading of plugin
options.

Combined, plugin options specified in sos.conf will now be logged
properly and this logging will occur before we set (and log the setting
of) those options.

Resolves: #2663

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>