[hostname_map] Add a catch for single-character hostnames
If a log file was truncated at a specific boundary in a string of the
FQDN of the host such that we only get a couple characters before the
rest of the domain, we would previously bodly replace all instances of
that character with the obfuscated short name; not very helpful.

Instead, don't sanitize the short name if this happens and instead
obfuscate the whole FQDN as 'unknown.example.com'.

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>