[origin] Removing unused configuration
Removed the diagnostics part as it is no longer maintained and doesn't
work on Openshift.

Adding additional projects to collect.

Removed getting all namespaces as it is not needed for troubleshooting
and project names are sensitive for some customers.

Adding condition to collect the logs from systemd openshift services if
not running as static pods.

Signed-off-by: Vladislav Walek <22072258+vwalek@users.noreply.github.com>