[collect] Add --transport option and allow clusters to set transport type
Adds a new `--transport` option for users to be able to specify the type
of transport to use when connecting to nodes. The default value of
`auto` will defer to the cluster profile to set the transport type,
which will continue to default to use OpenSSH's ControlPersist feature.

Clusters may override the new `set_transport_type()` method to change
the default transport used.

If `--transport` is anything besides `auto`, then the cluster profile
will not be deferred to when choosing a transport for each remote node.

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>