[Policy] Fix assignment of case id prompt
When prompting for a case id, `Policy` was not properly updating the
option value, only assigning the value to `Policy` which meant that
aspects outside of `Policy` could not always properly reference the
(updated) case id.

Fix this by assigning the case id prompt response back to the case_id
option value. `Policy` still retains a local reference to case_id as
existing logic was setting that based on the (assumed-to-be-updated)
option value, which until this commit would have been superfluous.

Closes: #2707

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>