[ocp] Create temporary project and restrict default node list to masters
Adds explicit setup of a new project to use in the `ocp` cluster and
adds better handling of cluster setup generally, which the `ocp` cluster
is the first to make use of.

Included in this change is a correction to
`Cluster.exec_primary_cmd()`'s use of `get_pty` to now be determined on
if the primary node is the local node or not.

Additionally, based on feedback from the OCP engineering team, by
default restrict node lists to masters.

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>