[Plugin] Allow writing command output directly to disk
This commit addresses a long standing ask in sos, regarding resource
consumption when sos is run with `--all-logs`.

For executions where `--all-logs` is used, or for specific commands
where `sizelimit` has been set to 0, `add_cmd_output()` and
`add_journal()` will now instruct `sos_get_command_output()` to write
output directly to a file rather than saving the output in memory.

When this occurs, the `output` key in the returned dict will be empty.

Note that this does extend to `collect_cmd_output()` or `exec_cmd()`.

Resolves: #1506

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>