[plugins] Update a few common plugins for new help output
Updates four common plugins to provide several examples of how to format
plugin documentation so that the new `help` component is able to provide
useful information on them.

Going forward, new plugins should include docstrings to provide this
help information as part of the initial review process. Existing plugins
will be updated as time allows or as plugins are updated and/or
refactored.

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>