[Plugin] Ensure specific plugin timeouts are only set for that plugin
It was discovered that setting a specific plugin timeout via the `-k
$plugin.timeout` option could influence the timeout setting for other
plugins that are not also having their timeout explicitly set. Fix this
by moving the default plugin opts into `Plugin.__init__()` so that each
plugin is ensured a private copy of these default plugin options.

Additionally, add more timeout data to plugin manifest entries to allow
for better tracking of this setting.

Adds a test case for this scenario.

Closes: #2744

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>