[sos] Reference TMPDIR environment variable and check fstype of tmpdir
If the `TMPDIR` env var is set, we should reference it if the user has
not provided `--tmp-dir` by the cmdline or sos.conf.

The order of precedence is now:

	1. cmdline use of `--tmp-dir`
	2. setting `tmp-dir` in `/etc/sos/sos.conf`
	3. the `TMPDIR` environment variable
	4. `/var/tmp` as a default

Additionally, we will now check if the filesystem type for our tmpdir is
tmpfs, and if so print a warning to the user about the potential
pitfalls of doing so. This information is now recorded in the manifest
as well.

Closes: #2738

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>