[report,collect] Add option to control default container runtime
Adds a new `--container-runtime` option that allows users to control
what default container runtime is used by plugins for container based
collections, effectively overriding policy defaults.

If no runtimes are active, this option is effectively ignored. If
however runtimes are active, but the requested one is not, raise an
exception to abort collection with an appropriate message to the user.

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>