[Plugin] Add container parameter for add_cmd_output()
Adds a new `container` parameter for `Plugin.add_cmd_output()`, which if
set will format all commands passed to that call for execution in the
specified container.

`Plugin.fmt_container_cmd()` is called for this purpose, and has been
modified so that if the given container does not exist, an empty string
is returned instead, thus preventing execution on the host.

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>