[networking] Put namespaced commands into subdirs
Where networking namespaces are used, there tend to be large numbers of
namespaces used. This in turn results in sos running and collecting very
large numbers of namespaced commands.

To aid in consumability, place these collections under a subdir for the
namespace under another "namespaces" subdir within the plugin directory.

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>