[systemd] Collect systemctl list-units --all
Systemd has a hard limit of units that it can keep, and the limit is 128K.

Once that limit is hit (which is usually happening due to various bugs),
systemd returns E2BIG ("Argument list too long") when trying to start a
new unit. Until recently, it was not possible to figure out what is
going on without looking at the "systemctl list-units --all" output.

That's why it's nice to have it in sos report.

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>