[report] Handle exceptionally old pexpect versions
Depending on configuration, certain downstreams may provide pexpect 2.3
or 4.6 (or later). The backport for SFTP upload support assumed a 4.x
version pexpect, however 2.x does not support the `encoding` parameter
to `pexpect.spawn()`.

Add a version check to determine if that parameter needs to be used or
not. Note that this does not need to be implemented against `main`, as
all supported downstreams for `main` support a minimum version of 4.x.

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>