[tests] Do not remove temp dirs from failed CI test runs
In the event a CI test execution fails before the archive is created,
for any reason, we want to preserve the temp directory as the logs there
are still useful.

As such, add a specific env var to our test runs via avocado. Then, when
we detect a failure check for that env var to determine if we should in
fact do our cleanup or not.

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>