[ovn_central] Account for Red Hat ovn package naming
Previous ovn packages were 'ovn2xxx' and now they have
been renamed to 'ovn-2xxx'. This causes sos tool to not
recognize that the packages are installed and it won't
collect the relevant data.

This patch is changing the match to be compatible
with the previous and newer naming conventions.

Signed-off-by: Daniel Alvarez Sanchez <dalvarez@redhat.com>