[cleaner] Improve concurrent archive cleaning performance
Prior to this change, concurrent archives were handled with a thread
pool. While this worked, the performance gains were not significant as
we were still locked into a single processor for all archives.

With this commit, we now use a process pool to obfuscate archives. `sos
clean` will now use up to `--jobs` processes (default 4) for concurrent
archive obfuscation. This results in a significant performance increase
for cleaning, for example, `sos collect` tarballs.

This requires moving several aspect of `SoSCleaner()` into different/new
objects so that the process pool can correctly pickle everything needed
in the new process space.

This in turn requires a shared memory object for the parser maps to
store their internal datasets. Multiprocessing dicts are inherently much
slower for access than standard dict objects, so in the event of single
archive obfuscations, parser maps retain the standard dict objects, and
we will only replace those with shared dicts for 2 or more concurrent
archives.

Signed-off-by: Jake Hunsaker <jhunsake@redhat.com>